Avoid resource contention and allow branches
- When we are running a deploy to staging, if more than one MR gets
merged in a short period of time we want to ensure that we are not
trying to write to the S3 bucket from multiple jobs at the same time.
- It was discovered that the change prevents the jobs that would have
run against a branch to deploy to a bucket got broken because it forced
the job to only run on master when previously setting a bucket would
have worked. This restores that functionality.