Enable a bump from external pipelines
This allows remote pipelines to bump a chart which will in turn make a change to master and kick off the normal pipeline.

Unfortunately, unlike the chatops bits, there is no way to easily tell the the external pipeline to only call a specific job (like we might call the publish-em or bump-chart jobs). I am open to other approaches that dont require annotating every job.