Fix handling of the csi driver and helmfile image
The CSI driver was being gated based on an env var that was not being threaded through. This corrects that, gates on cluster type instead since all of the saas clusters are 1.16. (The csi driver requires >= 1.17)

Also, this fixes the helmfile image. The maven changes hadn't been merged so fixed the image to include those changes.